pyproject.toml
[tool.poetry]
name = "openu-project-backend"
version = "0.1.0"
description = "Telegram Bot that used as money manager app"
authors = ["Sagiv Abu <sagivabu101@gmail.com>"]
readme = "README.md"
packages = [{include = "openu_project_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
matplotlib = "^3.7.1"
python-telegram-bot = "^20.2"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
